betsports/pyproject.toml
[project]
name = "betsports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.0",
    "django>=6.0",
    "django-baton>=5.0.3",
    "django-celery-beat>=2.1.0",
    "django-celery-results>=2.6.0",
    "django-filter>=25.2",
    "django-import-export>=4.3.14",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.12.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "pretty-html-table>=0.9.16",
    "psycopg[binary]>=3.3.2",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "ipython>=9.8.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "pytest-dotenv>=0.5.2",
    "pytest-xdist>=3.8.0",
    "werkzeug>=3.1.4",
]
